Adição do arquivo Cliente.cpp ao diretório src
