--- Cliente.cpp
@@ -1,13 +1,0 @@
-#include <iostream>
-#include <string>
-#include "Cliente.hpp"
-
-void Cliente::print(){
-
-	std::cout << "[Cliente]" << endl
-         << "  Nome: " << NOME << endl
-		 << "  Endereco: " << endereco << endl
-		 << "  CEP: " << Cep << endl;
-
-}
-
