+++ src/Cliente.cpp
@@ -1,0 +1,13 @@
+#include <iostream>
+#include <string>
+#include "Cliente.hpp"
+
+void Cliente::print(){
+
+	std::cout << "[Cliente]" << endl
+         << "  Nome: " << NOME << endl
+		 << "  Endereco: " << endereco << endl
+		 << "  CEP: " << Cep << endl;
+
+}
+
